+++ network.config.json
@@ -1,6 +1,6 @@
 {
   "xdcparentnet": "http://devnet.hashlabs.apothem.network:8545/",
-  "xdcsubnet": "http://66.94.98.186:8545/",
+  "xdcsubnet": "https://devnetstats.apothem.network/subnet",
   "xdcdevnet": "https://devnetstats.apothem.network/devnet",
   "xdctestnet": "https://erpc.apothem.network/"
 }
