+++ scripts/FullCheckpointDeploy.js
@@ -1,10 +1,17 @@
 const hre = require("hardhat");
 const deploy = require("../deployment.config.json");
-const subnet = require("./utils/subnet");
+const subnet = require("./utils/xdcnet");
+const network = require("../network.config.json");
+
 async function main() {
   const subnetDeploy = deploy["subnet"];
 
-  const { data0Encoded } = await subnet.data(subnetDeploy.gsbn);
+  const { data0Encoded } = await subnet.data(
+    network.xdcsubnet,
+    subnetDeploy.gsbn
+  );
+
+  console.log(data0Encoded);
 
   // We get the contract to deploy
   const checkpointFactory = await hre.ethers.getContractFactory(
